frame: Add AX25Path.
This represents the digipeater path taken (or to be taken) by a frame.
The CH bit is used to represent whether a frame actually has passed
through a given digipeater: the `reply` property gives the digipeaters
that have this CH bit set in reverse order so a reply message can be
provided.