frame: Add the ability to format a frame in TNC2 format
In `aprsd`, they're using a Python3 f-string to achieve this feature:
https://github.com/craigerl/aprsd/blob/master/aprsd/threads.py#L375

This works, but is kludgy because the `str` output for an `AX25Frame` is
meant to be a human-readable diagnostic output, whereas in this context
they're wanting a TNC2-style frame dump, which *is* meant to be read by
a machine.  Some of these "machine"s barf when they encounter something
unexpected:

https://github.com/sjlongland/aioax25/issues/9

APRS-IS mentions "TNC2 format" in its specs, but nowhere does it define
what is meant by "TNC2 format", so we're left to guess!  (If someone has
a definitive guide, without reading through the TAPR TNC2's firmware
source code, that'd be greatly appreciated!)