unit tests: Begin process of migrating to py.test
`nosetests` is no longer supported, so we need to shift.  Rather than
re-write all the tests to avoid pulling in `nose`, we'll just nick the
`nose.tools` assertions for now, with a view that we'll gradually port
tests over to the way they should be in `py.test` and eventually remove
this file.