github actions: Bump minimum tested version to 3.5+
I've tried to keep Python 3.4 in the loop, as it would enable quite a
few operating systems to continue to be used to run this library.
HOWEVER, with Python 3.10 making `asyncio.coroutine` deprecated, I
basically need to implement two parallel versions: a 3.4 version using
generator-based coroutines, and a modern 3.5+ version with `async def`.

With the removal of Python 3.4, the following platforms will also not
receive support:

- Microsoft Windows XP
- Debian GNU/Linux 8 (Jessie)