peer: Treat incoming SABM(E) in CONNECTING state as UA.
If we receive a `SABM(E)` while we are connecting to the target peer
station, treat this as a `UA` and send a `UA` in response.  This is not
strictly in the AX.25 specifications, but thinking about the situation,
seems the safest course of action to avoid getting into a "confused"
state.