unit: Add support for `pint` unit conversion
APRS was fundamentally designed in the US, and so uses imperial units
throughout its specification.  While I've got a great itch to metric-ise
the code base here, the truth is I'd have to re-work the equations which
were all written with imperial units in mind.

So, a half-way step: support unit conversion.  There's a good quality
library for doing this called `pint`, and so we'll optionally support
that.  Without that installed, we just assume the units that are native
to APRS itself (imperial).