station: Store and consider command bit state
This appears to be indicative of which side sent the SABM: sender sends
SABM with its own SSID field clearing the C/H bit, and the destination
SSID setting the C/H bit.

So in the station context, we really have two separate peers:
- Peer with C/H cleared: is a peer that sent US a `SABM(E)`
- Peer with C/H set: is a peer that WE sent a `SABM(E)`

Default to C/H set to `True`, since most users will be making outbound
connections.