futurequeue: Add handling of multiple futures
Where multiple callers call the same function, it is useful to just
queue the placeholders up and resolve or reject them all in one hit with
the same exception information.  This allows us to gracefully handle
multiple callers calling the same function and avoid concurrent multiple
execution of that function.