kiss: Use `FutureQueue` for managing open/close
This allows us to gracefully handle concurrent open/close requests.  If
`open()` is called while opening, and a future is provided, we enqueue
it.

Also, we add `return_future` to the constructor arguments, if this is
set to `True`; we always return a future if not provided one.  This
allows for programs to `await kissdev.open()` or
`await kissdev.close()`… whilst retaining the original one-shot
synchronous interface used by existing code.