github actions: Temporarily disable Python 3.5 for now
A certificate issue is preventing `pip` from installing packages.
Disable for now while we wait and see if the Github team can fix their
action or suggest changes to how it is used.

If they can't fix it (and lets face it, Python 3.5 is very ancient now,
it isn't in supported versions of Linux or *BSD), we'll bump the minimum
release to Python 3.6.

https://github.com/actions/setup-python/issues/866